product_clip_ai/requirements.txt
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
requests>=2.28.0
tqdm>=4.65.0

fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
